fix numpy dtype if np > 1.20

+++ geomdl/visualization/VisPlotly.py
@@ -58,7 +58,7 @@ class VisConfig(vis.VisConfigAbstract):
     """
     def __init__(self, **kwargs):
         super(VisConfig, self).__init__(**kwargs)
-        self.dtype = np.float
+        self.dtype = float
         # Set Plotly custom variables
         self.figure_image_filename = "temp-plot.html"
         self.use_renderer = kwargs.get("use_renderer", False)
